Move Python metadata function to its own file
